+++ lib/context.js
@@ -4,8 +4,6 @@
 
   nodes = require('./nodes');
 
-  IndexGenerator = require('./utils').IndexGenerator;
-
   exports.Name = Name = (function() {
     function Name(name1, type) {
       this.name = name1;
@@ -92,16 +90,17 @@
       var i, next;
       if (name === 'i') {
         next = this.indexGen.get();
-        while (this.has(next.name)) {
+        while (this.has(next)) {
           next = this.indexGen.get();
         }
-        this.push(new Name(next.name));
+        this.push(new Name(next));
         return next;
       } else {
         i = 0;
         while (true) {
           if (!this.has(name + (i === 0 ? '' : i))) {
-            return new nodes.Identifier(name + (i === 0 ? '' : i));
+            this.push(new Name(name + (i === 0 ? '' : i)));
+            return name + (i === 0 ? '' : i);
           }
           i++;
         }
@@ -109,6 +108,46 @@
     };
 
     return Context;
+
+  })();
+
+  exports.IndexGenerator = IndexGenerator = (function() {
+    function IndexGenerator() {}
+
+    IndexGenerator.prototype.letter = ['i', 'j', 'k', 'c', 'l', 'm', 'n', 'o', 'p', 'q', 'r', 's', 't', 'u', 'v', 'w', 'x', 'y', 'z', 'a', 'b', 'c', 'd', 'e', 'f', 'j', 'h'];
+
+    IndexGenerator.prototype._level = 0;
+
+    IndexGenerator.prototype.letterLevel = 0;
+
+    IndexGenerator.prototype.generateAtLevel = function() {
+      var i, r;
+      r = '';
+      i = 0;
+      while (i < this._level) {
+        r += '_';
+        i++;
+      }
+      return r;
+    };
+
+    IndexGenerator.prototype.next = function() {
+      if ((this.letterLevel + 1) === this.letter.length) {
+        this._level++;
+        return this.letterLevel = 0;
+      } else {
+        return this.letterLevel++;
+      }
+    };
+
+    IndexGenerator.prototype.get = function() {
+      var r;
+      r = this.generateAtLevel() + this.letter[this.letterLevel];
+      this.next();
+      return r;
+    };
+
+    return IndexGenerator;
 
   })();
