+++ lib/nodes.js
@@ -617,7 +617,7 @@
           return results2;
         })())).prepare(system).compile();
       } else {
-        index = system.context.free('i');
+        index = new Identifier(system.context.free('i'));
         expression = new Call(new Identifier('call_user_func'), [
           new Code([], new Block([
             new For({
@@ -1060,7 +1060,7 @@
         code = ' else ';
         code += this.body.prepare(system).compile(system);
       } else {
-        parentIf.closed = true;
+        this.parentIf.closed = true;
         code = this.body.prepare(system).compile(system);
       }
       return code;
@@ -1107,7 +1107,7 @@
       var cacheRes, funcgen;
       this.body.expands = true;
       if (this.returnactived) {
-        this.cacheRes = cacheRes = system.context.free('result');
+        this.cacheRes = cacheRes = new Identifier(system.context.free('result'));
         funcgen = function(exp) {
           var m;
           m = new Expression(new Call(new Value(new Identifier('array_push')), [new Value(cacheRes), exp]));
@@ -1189,7 +1189,7 @@
         code += this.TryBody.prepare(system).compile(system);
         code += ' catch(Exception ';
         if (this.CatchIdentifier === false) {
-          code += system.context.free('error').prepare(system).compile(system);
+          code += (new Identifier(system.context.free('error'))).prepare(system).compile(system);
         } else {
           code += this.CatchIdentifier.prepare(system).compile(system);
         }
@@ -1226,14 +1226,14 @@
       this.object = !!this.source.object;
       if (!((this.source.range != null) && this.source.range === true) && !this.isPrepared) {
         if (!this.object) {
-          this.cacheIndex = system.context.free('i');
-          this.cacheLen = system.context.free('len');
+          this.cacheIndex = new Identifier(system.context.free('i'));
+          this.cacheLen = new Identifier(system.context.free('len'));
           if (this.source.source.type === 'Value' && this.source.source.value.type === 'Identifier') {
             this.initRef = false;
             this.cacheRef = this.source.source.value;
           } else {
             this.initRef = true;
-            this.cacheRef = system.context.free('ref');
+            this.cacheRef = new Identifier(system.context.free('ref'));
           }
           valfromRef = new Value(this.cacheRef);
           valfromRef.add(new Access((this.source.index != null ? this.source.index : this.cacheIndex), '[]'));
@@ -1247,7 +1247,7 @@
         this.body.body.unshift(new Expression(new Assign('=', this.source.name, valfromRef)));
       }
       if (this.returnactived) {
-        this.cacheRes = cacheRes = system.context.free('result');
+        this.cacheRes = cacheRes = new Identifier(system.context.free('result'));
         funcgen = function(exp) {
           var m;
           m = new Expression(new Call(new Value(new Identifier('array_push')), [new Value(cacheRes), exp]));
@@ -1268,7 +1268,7 @@
           code += '\n' + system.indent.get();
         }
         if ((this.source.range != null) && this.source.range === true) {
-          index = this.source.index != null ? this.source.index : system.context.free('i');
+          index = this.source.index != null ? this.source.index : new Identifier(system.context.free('i'));
           code += 'for(';
           code += (new Assign('=', index, this.source.source.from)).prepare(system).compile(system);
           code += '; ';
@@ -1290,7 +1290,7 @@
             if (this.source.index != null) {
               code += this.source.index.prepare(system).compile(system);
             } else {
-              code += system.context.free('value').prepare(system).compile(system);
+              code += (new Identifier(system.context.free('value'))).prepare(system).compile(system);
             }
             code += ') ';
             code += this.body.prepare(system).compile(system);
