+++ lib/utils.js
@@ -1,6 +1,6 @@
 // Generated by CoffeeScript 1.9.3
 (function() {
-  var IndentGenerator, IndexGenerator, nodes;
+  var IndentGenerator, nodes;
 
   nodes = require('./nodes');
 
@@ -41,46 +41,6 @@
     };
 
     return IndentGenerator;
-
-  })();
-
-  exports.IndexGenerator = IndexGenerator = (function() {
-    function IndexGenerator() {}
-
-    IndexGenerator.prototype.letter = ['i', 'j', 'k', 'c', 'l', 'm', 'n', 'o', 'p', 'q', 'r', 's', 't', 'u', 'v', 'w', 'x', 'y', 'z', 'a', 'b', 'c', 'd', 'e', 'f', 'j', 'h'];
-
-    IndexGenerator.prototype._level = 0;
-
-    IndexGenerator.prototype.letterLevel = 0;
-
-    IndexGenerator.prototype.generateAtLevel = function() {
-      var i, r;
-      r = '';
-      i = 0;
-      while (i < this._level) {
-        r += '_';
-        i++;
-      }
-      return r;
-    };
-
-    IndexGenerator.prototype.next = function() {
-      if ((this.letterLevel + 1) === this.letter.length) {
-        this._level++;
-        return this.letterLevel = 0;
-      } else {
-        return this.letterLevel++;
-      }
-    };
-
-    IndexGenerator.prototype.get = function() {
-      var r;
-      r = this.generateAtLevel() + this.letter[this.letterLevel];
-      this.next();
-      return new nodes.Identifier(r);
-    };
-
-    return IndexGenerator;
 
   })();
